split lib.rs into distinct modules
