+++ src/lib.rs
@@ -1,66 +1,23 @@
-pub mod codec;
+pub mod client;
+pub use client::RPCClient;
+pub use client::RPCRequest;
+
+pub mod stream;
+pub use stream::StreamClient;
+pub use stream::StreamUpdate;
+
 pub mod error;
+
+// Re-exported for the generated code
+pub mod codec;
 pub mod krpc;
-use error::Error;
-use error::Result;
-
 pub use protobuf;
-use protobuf::Message;
-
-use std::collections::HashMap;
-use std::marker::PhantomData;
-use std::net::TcpStream;
-use std::net::ToSocketAddrs;
-
-/// A client to the RPC server.
-#[derive(Debug)]
-pub struct RPCClient {
-    sock: TcpStream,
-    client_id: Vec<u8>,
-}
-
-/// A client to the Stream server.
-#[derive(Debug)]
-pub struct StreamClient {
-    sock: TcpStream,
-}
-
-type StreamID = u64;
-
-/// Represents a request that can be submitted to the RPCServer. This object is clonable so that
-/// you can perform the same request multiple times. For one-off requests, use [`batch_call`],
-/// [`batch_call_unwrap`] or [`RPCClient::mk_call`] which provide a nicer API.
-#[derive(Clone, Default)]
-pub struct RPCRequest {
-    calls: protobuf::RepeatedField<krpc::ProcedureCall>,
-}
-
-/// Represents a procedure call. The type parameter is the type of the value to be extracted from
-/// the server's response.
-#[derive(Clone)]
-pub struct CallHandle<T> {
-    proc_call: krpc::ProcedureCall,
-    _phantom: PhantomData<T>,
-}
-
-/// A handle to a stream. The type parameter is the type of the value produced by the stream.
-#[derive(Copy, Clone, Debug)]
-pub struct StreamHandle<T> {
-    stream_id: StreamID,
-    _phantom: PhantomData<T>,
-}
-
-/// A collection of updates received from the stream server.
-#[derive(Debug, Clone, Default)]
-pub struct StreamUpdate {
-    updates: HashMap<StreamID, krpc::ProcedureResult>,
-}
 
 #[doc(hidden)]
 #[macro_export]
 macro_rules! batch_call_common {
     ($process_result:expr, $client:expr, ( $( $call:expr ),+ )) => {{
-        let mut request = $crate::RPCRequest::default();
+        let mut request = $crate::client::RPCRequest::default();
         $( request.add_call($call); )+
         match $client.submit_request(request) {
             Err(e) => {
@@ -97,196 +54,10 @@ macro_rules! batch_call {
     };
 }
 
-/// Does the same as [`batch_call`] but unwraps all values automatically.
+/// Does the same as [`batch_call!`] but unwraps all values automatically.
 #[macro_export]
 macro_rules! batch_call_unwrap {
     ($client:expr, ( $( $call:expr ),+ $(,)? )) => {{
         $crate::batch_call_common!(|result: $crate::error::Result<_>| { result.unwrap() }, $client, ( $( $call ),+ ))
     }};
-}
-
-/// Creates a stream request from a CallHandle. For less verbosity, you can use the
-/// [`CallHandle::to_stream`] instead.
-///
-/// Note that types don't prevent you from chaining multiple `mk_stream`. This will build a stream
-/// request of a stream request. Turns out this is accepted by the RPC server and the author of
-/// this library confesses he had some fun with this.
-pub fn mk_stream<T: codec::RPCExtractable>(call: &CallHandle<T>) -> CallHandle<StreamHandle<T>> {
-    let mut arg = krpc::Argument::new();
-    arg.set_position(0);
-    arg.set_value(call.get_call().write_to_bytes().unwrap());
-
-    let mut arguments = protobuf::RepeatedField::<krpc::Argument>::new();
-    arguments.push(arg);
-
-    let mut proc_call = krpc::ProcedureCall::new();
-    proc_call.set_service(String::from("KRPC"));
-    proc_call.set_procedure(String::from("AddStream"));
-    proc_call.set_arguments(arguments);
-
-    CallHandle::<StreamHandle<T>>::new(proc_call)
-}
-
-impl RPCClient {
-    /// Connects to the KRPC server. The client will show up in the KRPC UI with the given client name.
-    pub fn connect<A: ToSocketAddrs>(client_name: &str, addr: A) -> Result<Self> {
-        let mut sock = TcpStream::connect(addr)?;
-
-        let mut conn_req = krpc::ConnectionRequest::new();
-        conn_req.set_field_type(krpc::ConnectionRequest_Type::RPC);
-        conn_req.set_client_name(client_name.to_string());
-
-        conn_req.write_length_delimited_to_writer(&mut sock)?;
-
-        let mut response = codec::read_message::<krpc::ConnectionResponse>(&mut sock)?;
-
-        match response.status {
-            krpc::ConnectionResponse_Status::OK => Ok(RPCClient {
-                sock,
-                client_id: response.client_identifier,
-            }),
-            s => Err(Error::RPCConnect {
-                error: response.take_message(),
-                status: s,
-            }),
-        }
-    }
-
-    /// Sends a single RPC request to the server.
-    pub fn mk_call<T: codec::RPCExtractable>(&mut self, call: &CallHandle<T>) -> Result<T> {
-        let (result,) = batch_call!(self, (call))?;
-        result
-    }
-
-    /// Sends an [`RPCRequest`] to the server. A single RPCRequest may contain multiple RPC calls.
-    /// It is recommended to use the [`batch_call`] or [`batch_call_unwrap`] for one-off
-    /// requests.
-    pub fn submit_request(&mut self, request: RPCRequest) -> Result<krpc::Response> {
-        let raw_request = request.build();
-        raw_request.write_length_delimited_to_writer(&mut self.sock)?;
-        let resp = codec::read_message::<krpc::Response>(&mut self.sock)?;
-        Ok(resp)
-    }
-}
-
-impl StreamClient {
-    pub fn connect<A: ToSocketAddrs>(client: &RPCClient, addr: A) -> Result<Self> {
-        let mut sock = TcpStream::connect(addr)?;
-
-        let mut conn_req = krpc::ConnectionRequest::new();
-        conn_req.set_field_type(krpc::ConnectionRequest_Type::STREAM);
-        conn_req.set_client_identifier(client.client_id.clone());
-
-        conn_req.write_length_delimited_to_writer(&mut sock)?;
-
-        let mut response = codec::read_message::<krpc::ConnectionResponse>(&mut sock)?;
-
-        match response.status {
-            krpc::ConnectionResponse_Status::OK => Ok(Self { sock }),
-            s => Err(Error::StreamConnect {
-                error: response.take_message(),
-                status: s,
-            }),
-        }
-    }
-
-    pub fn recv_update(&mut self) -> Result<StreamUpdate> {
-        let updates = codec::read_message::<krpc::StreamUpdate>(&mut self.sock)?;
-
-        let mut map = HashMap::new();
-        for mut result in updates.results.into_iter() {
-            map.insert(result.id, result.take_result());
-        }
-
-        Ok(StreamUpdate { updates: map })
-    }
-}
-
-impl RPCRequest {
-    pub fn add_call<T: codec::RPCExtractable>(&mut self, handle: &CallHandle<T>) {
-        self.calls.push(handle.get_call().clone())
-    }
-
-    fn build(self) -> krpc::Request {
-        let mut req = krpc::Request::new();
-        req.set_calls(self.calls);
-        req
-    }
-}
-
-impl<T> CallHandle<T>
-where
-    T: codec::RPCExtractable,
-{
-    #[doc(hidden)]
-    /// Creates a new CallHandle. The function is public so that the generated code from
-    /// krpc-mars-terraformer can use it but it is hidden from user docs.
-    pub fn new(proc_call: krpc::ProcedureCall) -> Self {
-        CallHandle {
-            proc_call,
-            _phantom: PhantomData,
-        }
-    }
-
-    pub fn to_stream(&self) -> CallHandle<StreamHandle<T>> {
-        mk_stream(self)
-    }
-
-    pub fn get_result(&self, result: &krpc::ProcedureResult) -> Result<T> {
-        codec::extract_result(result)
-    }
-
-    fn get_call(&self) -> &krpc::ProcedureCall {
-        &self.proc_call
-    }
-}
-
-impl<T> StreamHandle<T> {
-    #[doc(hidden)]
-    /// Creates a new StreamHande. The function is public so that the generated code from
-    /// krpc-mars-terraformer can use it but it is hidden from user docs.
-    pub fn new(stream_id: StreamID) -> Self {
-        StreamHandle {
-            stream_id,
-            _phantom: PhantomData,
-        }
-    }
-
-    /// Creates an RPC request that will remove this stream.
-    pub fn remove(self) -> CallHandle<()> {
-        use codec::RPCEncodable;
-
-        let mut arg = krpc::Argument::new();
-        arg.set_position(0);
-        arg.set_value(self.stream_id.encode_to_bytes().unwrap());
-
-        let mut arguments = protobuf::RepeatedField::<krpc::Argument>::new();
-        arguments.push(arg);
-
-        let mut proc_call = krpc::ProcedureCall::new();
-        proc_call.set_service(String::from("KRPC"));
-        proc_call.set_procedure(String::from("RemoveStream"));
-        proc_call.set_arguments(arguments);
-
-        CallHandle::<()>::new(proc_call)
-    }
-}
-
-impl StreamUpdate {
-    pub fn get_result<T>(&self, handle: &StreamHandle<T>) -> Result<T>
-    where
-        T: codec::RPCExtractable,
-    {
-        let result = self
-            .updates
-            .get(&handle.stream_id)
-            .ok_or(Error::NoSuchStream)?;
-        codec::extract_result(&result)
-    }
-
-    /// Merge two update objects. The Stream server doesn't update values that don't change, so
-    /// this can be used to retain previous values of streams.
-    pub fn merge_with(&mut self, other: StreamUpdate) {
-        self.updates.extend(other.updates)
-    }
 }
